:white_check_mark: add message tests
